+++ docs/changelog.md
@@ -3,6 +3,7 @@
 ## 1.2 (unreleased)
 
 * Quote commodities containing minus/dash (`-`) ([issue #15](https://github.com/beancount/beancount2ledger/issues/15))
+* Turn multi-line strings into a single line ([issue #17](https://github.com/beancount/beancount2ledger/issues/17))
 * Fix crash with very long account names ([issue #19](https://github.com/beancount/beancount2ledger/issues/19))
 * Fix crash when entry.meta or entry.posting are not set ([issue #16](https://github.com/beancount/beancount2ledger/issues/16))